@@ -1,77 +1,61 @@
 {
   "name": "contacts-app-starter",
   "version": "1.0.0",
-  "license": "MIT",
-  "engines": {
-    "node": ">=6.9.0"
-  },
-  "angular-cli": {},
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
+    "start": "ng serve -o",
     "build": "ng build",
     "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "protractor",
-    "karma": "karma",
-    "webdriver-manager": "webdriver-manager"
+    "e2e": "protractor"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.1.1",
-    "@angular/common": "5.1.1",
-    "@angular/compiler": "5.1.1",
-    "@angular/core": "5.1.1",
-    "@angular/forms": "5.1.1",
-    "@angular/http": "5.1.1",
-    "@angular/material": "5.0.2",
-    "@angular/platform-browser": "5.1.1",
-    "@angular/platform-browser-dynamic": "5.1.1",
-    "@angular/platform-server": "5.1.1",
-    "@angular/router": "5.1.1",
-    "angular-in-memory-web-api": "0.5.2",
-    "core-js": "2.5.3",
+    "@angular/animations": "15.2.6",
+    "@angular/cdk": "^15.2.5",
+    "@angular/common": "15.2.6",
+    "@angular/compiler": "15.2.6",
+    "@angular/core": "15.2.6",
+    "@angular/forms": "15.2.6",
+    "@angular/material": "^15.2.5",
+    "@angular/platform-browser": "15.2.6",
+    "@angular/platform-browser-dynamic": "15.2.6",
+    "@angular/platform-server": "15.2.6",
+    "@angular/router": "15.2.6",
+    "angular-in-memory-web-api": "0.15.0",
+    "core-js": "3.30.0",
     "hammerjs": "2.0.8",
-    "rxjs": "5.5.5",
-    "web-animations-js": "2.3.1",
-    "zone.js": "0.8.18"
+    "rxjs": "7.8.0",
+    "tslib": "^2.3.0",
+    "web-animations-js": "2.3.2",
+    "zone.js": "0.13.0"
   },
   "devDependencies": {
-    "@angular/cdk": "5.0.2",
-    "@angular/cli": "1.6.8",
-    "@angular/compiler-cli": "5.1.1",
-    "@angular/language-service": "5.1.1",
-    "@types/hammerjs": "2.0.35",
-    "@types/jasmine": "2.8.2",
-    "@types/jasminewd2": "2.0.3",
-    "@types/node": "8.5.1",
-    "codelyzer": "4.0.2",
-    "jasmine-core": "2.8.0",
-    "jasmine-reporters": "2.2.1",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "1.7.1",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "1.3.0",
-    "karma-jasmine": "1.1.1",
-    "karma-jasmine-html-reporter": "0.2.2",
-    "karma-junit-reporter": "1.2.0",
-    "protractor": "5.2.2",
-    "ts-node": "4.0.2",
-    "tslint": "5.8.0",
-    "typescript": "2.4.2"
+    "@angular-devkit/build-angular": "^15.2.5",
+    "@angular/cdk": "^15.2.5",
+    "@angular/cli": "^15.2.5",
+    "@angular/compiler-cli": "15.2.6",
+    "@angular/language-service": "15.2.6",
+    "@types/hammerjs": "2.0.41",
+    "@types/jasmine": "~4.3.0",
+    "@types/jasminewd2": "2.0.10",
+    "@types/node": "18.15.11",
+    "codelyzer": "6.0.2",
+    "jasmine-core": "~4.5.0",
+    "jasmine-reporters": "2.5.2",
+    "jasmine-spec-reporter": "7.0.0",
+    "karma": "6.4.1",
+    "karma-chrome-launcher": "3.1.1",
+    "karma-cli": "2.0.0",
+    "karma-coverage-istanbul-reporter": "3.0.3",
+    "karma-jasmine": "5.1.0",
+    "karma-jasmine-html-reporter": "~2.0.0",
+    "karma-junit-reporter": "2.0.1",
+    "protractor": "7.0.0",
+    "ts-node": "10.9.1",
+    "tslint": "6.1.3",
+    "typescript": "4.9.5"
   },
   "repository": {
     "url": "git+https://github.com/testing-angular-applications/testing-angular-applications.git"
-  },
-  "bugs": {
-    "url": "https://github.com/testing-angular-applications/testing-angular-applications/issues"
-  },
-  "homepage": "https://github.com/testing-angular-applications/testing-angular-applications#readme",
-  "contributors": [
-    "Jesse Palmer <jesselpalmer@gmail.com>",
-    "Craig Nishina <craig.nishina@gmail.com>",
-    "Michael Giambalvo <heathkit@gmail.com>",
-    "Corinna Cohn <corinna.cohn@gmail.com>"
-  ]
+  }
 }
