@@ -1,13 +1,11 @@
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 import { InMemoryWebApiModule } from 'angular-in-memory-web-api';
 import { InMemoryDataService } from './in-memory-data.service';
-
-import './rxjs-extensions';
 
 import { AppMaterialModule } from './app.material.module';
 import { ContactDetailComponent } from './contacts/contact-detail';
@@ -54,7 +52,7 @@ import { InvalidPhoneNumberModalComponent } from './contacts/shared';
     BrowserModule,
     BrowserAnimationsModule,
     FormsModule,
-    HttpModule,
+    HttpClientModule,
     InMemoryWebApiModule.forRoot(InMemoryDataService, { dataEncapsulation: true }),
     RoutingModule
   ],
